requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
requests
python-dotenv